pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "June P", email = "rjp2030@gmail.com"},
]
dependencies = [
    "discord-py-self>=2.0.0",
    "sqlalchemy>=2.0.20",
    "cython>=3.0.0",
    "nuitka>=1.7.10",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "AGPL-3.0-Or-Later"}

